requirements.txt
blinker==1.4
click==7.1.2
dnspython==1.16.0
Flask==1.1.4
itsdangerous==0.24
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.4
python-dotenv==0.15.0
Werkzeug==1.0.1
